Tests: Split buildiso tests into submodule
