+++ tests/actions/buildiso/append_line_test.py
@@ -1,0 +1,56 @@
+import pytest
+
+from cobbler.actions import mkloaders
+from cobbler.actions.buildiso.netboot import AppendLineBuilder
+from cobbler import utils
+
+
+@pytest.fixture(scope="function", autouse=True)
+def create_loaders(cobbler_api):
+    loaders = mkloaders.MkLoaders(cobbler_api)
+    loaders.run()
+
+
+def test_init():
+    assert isinstance(AppendLineBuilder("", {}), AppendLineBuilder)
+
+
+def test_generate_system(cobbler_api, create_distro, create_profile, create_system):
+    # Arrange
+    test_distro = create_distro()
+    test_distro.breed = "suse"
+    cobbler_api.add_distro(test_distro)
+    test_profile = create_profile(test_distro.name)
+    test_system = create_system(profile_name=test_profile.name)
+    blendered_data = utils.blender(cobbler_api, False, test_system)
+    test_builder = AppendLineBuilder(test_distro.name, blendered_data)
+
+    # Act
+    result = test_builder.generate_system(test_distro, test_system, False)
+
+    # Assert
+    # Very basic test yes but this is the expected result atm
+    # TODO: Make tests more sophisticated
+    assert (
+        result
+        == "  APPEND initrd=testdistro.img install=http://192.168.1.1:80/cblr/links/testdistro autoyast=default.ks"
+    )
+
+
+def test_generate_profile(cobbler_api, create_distro, create_profile):
+    # Arrange
+    test_distro = create_distro()
+    test_profile = create_profile(test_distro.name)
+    blendered_data = utils.blender(cobbler_api, False, test_profile)
+    test_builder = AppendLineBuilder(test_distro.name, blendered_data)
+
+    # Act
+    result = test_builder.generate_profile("suse")
+
+    # Assert
+    # Very basic test yes but this is the expected result atm
+    # TODO: Make tests more sophisticated
+    assert (
+        result
+        == " append initrd=testdistro.img install=http://192.168.1.1:80/cblr/links/testdistro autoyast=default.ks"
+    )
